Moved things from util that should be in input.dbms
