+++ src/main/java/org/schemaspy/input/dbms/ConnectionURLBuilder.java
@@ -20,9 +20,11 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.schemaspy.util;
+package org.schemaspy.input.dbms;
 
 import org.schemaspy.Config;
+import org.schemaspy.util.DbSpecificConfig;
+import org.schemaspy.util.DbSpecificOption;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
