+++ src/main/java/org/schemaspy/input/dbms/service/DatabaseService.java
@@ -29,7 +29,6 @@ import org.schemaspy.input.dbms.xml.SchemaMeta;
 import org.schemaspy.input.dbms.xml.TableMeta;
 import org.schemaspy.model.*;
 import org.schemaspy.service.helper.RemoteTableIdentifier;
-import org.schemaspy.util.DurationFormatter;
 import org.schemaspy.validator.NameValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
